fix(suite-native): T1 pin loading on passphrase
